Also move mock sequencer, rate limiter, terminination predicate
