+++ include/nighthawk/common/BUILD
@@ -27,9 +27,10 @@ envoy_basic_cc_library(
     include_prefix = "nighthawk/common",
     deps = [
         ":request_lib",
+        "//api/client:base_cc_proto",
         "@envoy//source/common/common:minimal_logger_lib",
-        "@envoy//source/common/common:non_copyable",
-        "@envoy//source/common/event:dispatcher_lib",
+        "@envoy//source/common/common:non_copyable_with_external_headers",
+        "@envoy//source/common/event:dispatcher_lib_with_external_headers",
         "@envoy//source/common/network:utility_lib_with_external_headers",
     ],
 )
