+++ test/client_worker_test.cc
@@ -18,6 +18,8 @@
 #include "test/mocks.h"
 #include "test/mocks/client/mock_benchmark_client.h"
 #include "test/mocks/client/mock_options.h"
+#include "test/mocks/common/mock_sequencer.h"
+#include "test/mocks/common/mock_termination_predicate.h"
 
 #include "gtest/gtest.h"
