+++ test/mocks.cc
@@ -1,18 +1,10 @@
 #include "test/mocks.h"
 
-#include <memory>
-
 #include "gmock/gmock.h"
 
 namespace Nighthawk {
 
 MockPlatformUtil::MockPlatformUtil() = default;
-
-MockRateLimiter::MockRateLimiter() = default;
-
-MockSequencerTarget::MockSequencerTarget() = default;
-
-MockSequencer::MockSequencer() = default;
 
 MockBenchmarkClientFactory::MockBenchmarkClientFactory() = default;
 
@@ -27,9 +19,5 @@ MockRequestSourceFactory::MockRequestSourceFactory() = default;
 MockTerminationPredicateFactory::MockTerminationPredicateFactory() = default;
 
 MockRequestSource::MockRequestSource() = default;
-
-MockTerminationPredicate::MockTerminationPredicate() = default;
-
-MockDiscreteNumericDistributionSampler::MockDiscreteNumericDistributionSampler() = default;
 
 } // namespace Nighthawk
