+++ test/mocks.h
@@ -11,11 +11,8 @@
 
 #include "nighthawk/client/factories.h"
 #include "nighthawk/common/platform_util.h"
-#include "nighthawk/common/rate_limiter.h"
 #include "nighthawk/common/request_source.h"
-#include "nighthawk/common/sequencer.h"
 #include "nighthawk/common/statistic.h"
-#include "nighthawk/common/termination_predicate.h"
 #include "nighthawk/common/uri.h"
 
 #include "external/envoy/test/test_common/simulated_time_system.h"
@@ -37,28 +34,6 @@ public:
 
   MOCK_CONST_METHOD0(yieldCurrentThread, void());
   MOCK_CONST_METHOD1(sleep, void(std::chrono::microseconds));
-};
-
-class MockRateLimiter : public RateLimiter {
-public:
-  MockRateLimiter();
-
-  MOCK_METHOD0(tryAcquireOne, bool());
-  MOCK_METHOD0(releaseOne, void());
-  MOCK_METHOD0(timeSource, Envoy::TimeSource&());
-  MOCK_METHOD0(elapsed, std::chrono::nanoseconds());
-};
-
-class MockSequencer : public Sequencer {
-public:
-  MockSequencer();
-
-  MOCK_METHOD0(start, void());
-  MOCK_METHOD0(waitForCompletion, void());
-  MOCK_CONST_METHOD0(completionsPerSecond, double());
-  MOCK_CONST_METHOD0(executionDuration, std::chrono::nanoseconds());
-  MOCK_CONST_METHOD0(statistics, StatisticPtrMap());
-  MOCK_METHOD0(cancel, void());
 };
 
 class MockBenchmarkClientFactory : public Client::BenchmarkClientFactory {
@@ -114,41 +89,11 @@ public:
                                              const Envoy::MonotonicTime scheduled_starting_time));
 };
 
-class FakeSequencerTarget {
-public:
-  virtual ~FakeSequencerTarget() = default;
-  // A fake method that matches the sequencer target signature.
-  virtual bool callback(OperationCallback) PURE;
-};
-
-class MockSequencerTarget : public FakeSequencerTarget {
-public:
-  MockSequencerTarget();
-  MOCK_METHOD1(callback, bool(OperationCallback));
-};
-
 class MockRequestSource : public RequestSource {
 public:
   MockRequestSource();
   MOCK_METHOD0(get, RequestGenerator());
   MOCK_METHOD0(initOnThread, void());
-};
-
-class MockTerminationPredicate : public TerminationPredicate {
-public:
-  MockTerminationPredicate();
-  MOCK_METHOD1(link, TerminationPredicate&(TerminationPredicatePtr&&));
-  MOCK_METHOD1(appendToChain, TerminationPredicate&(TerminationPredicatePtr&&));
-  MOCK_METHOD0(evaluateChain, TerminationPredicate::Status());
-  MOCK_METHOD0(evaluate, TerminationPredicate::Status());
-};
-
-class MockDiscreteNumericDistributionSampler : public DiscreteNumericDistributionSampler {
-public:
-  MockDiscreteNumericDistributionSampler();
-  MOCK_METHOD0(getValue, uint64_t());
-  MOCK_CONST_METHOD0(min, uint64_t());
-  MOCK_CONST_METHOD0(max, uint64_t());
 };
 
 } // namespace Nighthawk
