+++ test/mocks/common/BUILD
@@ -1,0 +1,39 @@
+licenses(["notice"])  # Apache 2
+
+load(
+    "@envoy//bazel:envoy_build_system.bzl",
+    "envoy_cc_mock",
+    "envoy_package",
+)
+
+envoy_package()
+
+envoy_cc_mock(
+    name = "mock_rate_limiter",
+    srcs = ["mock_rate_limiter.cc"],
+    hdrs = ["mock_rate_limiter.h"],
+    repository = "@envoy",
+    deps = [
+        "//include/nighthawk/common:base_includes",
+    ],
+)
+
+envoy_cc_mock(
+    name = "mock_sequencer",
+    srcs = ["mock_sequencer.cc"],
+    hdrs = ["mock_sequencer.h"],
+    repository = "@envoy",
+    deps = [
+        "//include/nighthawk/common:base_includes",
+    ],
+)
+
+envoy_cc_mock(
+    name = "mock_termination_predicate",
+    srcs = ["mock_termination_predicate.cc"],
+    hdrs = ["mock_termination_predicate.h"],
+    repository = "@envoy",
+    deps = [
+        "//include/nighthawk/common:base_includes",
+    ],
+)
