+++ test/mocks/common/mock_rate_limiter.cc
@@ -1,0 +1,9 @@
+#include "test/mocks/common/mock_rate_limiter.h"
+
+namespace Nighthawk {
+
+MockRateLimiter::MockRateLimiter() = default;
+
+MockDiscreteNumericDistributionSampler::MockDiscreteNumericDistributionSampler() = default;
+
+} // namespace Nighthawk
