+++ test/mocks/common/mock_sequencer.cc
@@ -1,0 +1,7 @@
+#include "test/mocks/common/mock_sequencer.h"
+
+namespace Nighthawk {
+
+MockSequencer::MockSequencer() = default;
+
+}
