+++ test/mocks/common/mock_termination_predicate.cc
@@ -1,0 +1,7 @@
+#include "test/mocks/common/mock_termination_predicate.h"
+
+namespace Nighthawk {
+
+MockTerminationPredicate::MockTerminationPredicate() = default;
+
+}
